rename precondition file
https://coveord.atlassian.net/browse/CDX-279
