+++ packages/cli/src/lib/decorators/preconditions/index.ts
@@ -2,5 +2,5 @@ export * from './node';
 export * from './npx';
 export * from './npm';
 export * from './authenticated';
-export * from './impersonatePrivilege';
+export * from './apiKeyPrivilege';
 export * from './preconditions';
